Ignore all Grunt warnings
Unfortunately `grunt-contrib-compass`, the task runner plugin used to
run `compass` (the Ruby dependency), requires version >=0.12.2 of
`compass` but installing that recent of a version of `compass` requires
Ruby >=2.0.0 which is a major revision bump which certainly includes
breaking changes. Thus we ignore all warnings and let CSS be broken so
that the server can at least run.