Fix auto-complete
There were two issues going on here:

1. The ArcGIS geocoder seems to now require https and we were using http
2. When the mapping service starts up, `buildingIds` is populated with
   data from CartoDB, specifically Philadelphia building ids to be used
for suggestions. The issue was some of the building ids returned by
CartoDB could be null and thus trying to access a property of null would
throw a type error which was then swallowed up by Angular, resulting in
a rejected promise and nothing happening in the UI. It seems that the
code in `searchBuildingIds` was counting on the building ids being
strings to make use of `String.prototype.indexOf`, though all the
building ids returned are currently numbers so it's unclear when this
last worked.