Fix autocomplete (#282)
* Attempt to Dockerize project

This is running into issues building because compass requires Ruby >= 2.0.0

* Force install of Ruby package to avoid dependency requirements

Also ignore `package-lock.json` since without that running the container
would fail with this error:

grunt[1]: ../src/node_contextify.cc:635:static void
node::contextify::ContextifyScript::New(const
v8::FunctionCallbackInfo<v8::Value>&): Assertion `args[1]->IsString()'
failed.
 1: 0x8fb090 node::Abort() [grunt]
 2: 0x8fb165  [grunt]
 3: 0x92fa6e
    node::contextify::ContextifyScript::New(v8::FunctionCallbackInfo<v8::Value>
const&) [grunt]
 4: 0xb903db  [grunt]
 5: 0xb92372 v8::internal::Builtin_HandleApiCall(int,
    v8::internal::Object**, v8::internal::Isolate*) [grunt]
 6: 0x370115a5bf1d

Avoiding building `package-lock.json` into the image fixed the issue as
per this comment:
https://github.com/nodejs/node/issues/20325#issuecomment-462421059

* Upgrade Node to 10 so deps can install successfully

The older version of Node that was running in the container was giving
cert issues.

* Ignore all Grunt warnings

Unfortunately `grunt-contrib-compass`, the task runner plugin used to
run `compass` (the Ruby dependency), requires version >=0.12.2 of
`compass` but installing that recent of a version of `compass` requires
Ruby >=2.0.0 which is a major revision bump which certainly includes
breaking changes. Thus we ignore all warnings and let CSS be broken so
that the server can at least run.

* Fix auto-complete

There were two issues going on here:

1. The ArcGIS geocoder seems to now require https and we were using http
2. When the mapping service starts up, `buildingIds` is populated with
   data from CartoDB, specifically Philadelphia building ids to be used
for suggestions. The issue was some of the building ids returned by
CartoDB could be null and thus trying to access a property of null would
throw a type error which was then swallowed up by Angular, resulting in
a rejected promise and nothing happening in the UI. It seems that the
code in `searchBuildingIds` was counting on the building ids being
strings to make use of `String.prototype.indexOf`, though all the
building ids returned are currently numbers so it's unclear when this
last worked.

* Upgrade to Xenial, Ruby 2.3, and install all Ruby deps without forcing

This should allow for deploying to production

* Don't ignore Grunt warnings anymore

since it can be served without warnings.

* Update README

* Use updated package-lock.json

This file was updated when building the image

* Attempt to fix build

Fixing the build and deploying to staging through CI would be good to build confidence. This copies the changes in the Dockerfile over to

* Remove CMD from Dockerfile

* Update README.md

Co-authored-by: Derek Dohler <ddohler@azavea.com>

Co-authored-by: Derek Dohler <ddohler@azavea.com>